separate patch code from main pybump logic
