+++ .github/workflows/pythonpackage.yml
@@ -30,7 +30,7 @@ jobs:
 
     - name: Lint with flake8
       run: |
-        pip install flake8
+        pip install flake8==5.0.4
         # stop the build if there are Python syntax errors or undefined names
         flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics --exclude=venv
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
